hand_tracked_musical_interface/requirements.txt
mediapipe
opencv-python
mido
python-rtmidi
PyYAML
pygame
